% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportita-package.R
\docType{package}
\name{apportita-package}
\alias{apportita}
\alias{apportita-package}
\title{apportita: Utility for Handling 'magnitude' Word Embeddings}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A partial R port from 'magnitude', which is a fast, simple utility library for handling vector embeddings. The main goal of this package is to enable access to user's local magnitude data store.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
