% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distances from keys to keys in q}
\usage{
calc_dist(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming"),
  ...
)
}
\arguments{
\item{con}{a Magnitude connection.}

\item{key}{character vector.}

\item{q}{character vector.}

\item{normalized}{logical.}

\item{...}{other arguments are passed to \code{proxyC::dist}.}

\item{mehtod}{string; method to compute distance.}
}
\value{
a sparse Matrix of 'Matrix' package.
}
\description{
Calculate distances from keys to keys in q
}
