% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-wrd.R
\name{calc_wrd}
\alias{calc_wrd}
\title{Calculate Word Rotator's Distance from keys to keys}
\usage{
calc_wrd(conn, keys, q, normalized = TRUE, ...)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{keys}{character vector.}

\item{q}{character vector.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{...}{other arguments are passed to \code{transport::wasserstein} internally.}
}
\value{
numeric scalar.
}
\description{
Calculate Word Rotator's Distance from keys to keys
}
