% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{close,Magnitude-method}
\alias{close,Magnitude-method}
\title{Close a Magnitude connection}
\usage{
\S4method{close}{Magnitude}(con)
}
\arguments{
\item{con}{a Magnitude connection.}
}
\value{
the value from \code{RSQLite::dbDisconnect} is returned invisibly.
}
\description{
Close a Magnitude connection
}
