% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{dim,Magnitude-method}
\alias{dim,Magnitude-method}
\title{Rows size of a Magnitude table
Columns size of a Magnitude table
Dimensions of a Magnitude table}
\usage{
\S4method{dim}{Magnitude}(x)
}
\description{
Rows size of a Magnitude table
Columns size of a Magnitude table
Dimensions of a Magnitude table
}
