% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{doesnt_match}
\alias{doesnt_match}
\title{Order q by their distances to key}
\usage{
doesnt_match(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming"),
  n = 1L
)
}
\arguments{
\item{con}{a Magnitude connection.}

\item{key}{string.}

\item{q}{character vector.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{method}{string; method to compute distance.}

\item{n}{integer.}
}
\value{
an ordered named numeric vector of which elements represent distances to `key`.
}
\description{
Order q by their distances to key
}
