% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{has_exact}
\alias{has_exact}
\title{Check if keys exist in a Magnitude table?}
\usage{
has_exact(conn, keys)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{keys}{a character vector.}
}
\value{
a tibble.
}
\description{
Check if keys exist in a Magnitude table?
}
