% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude-class}
\alias{magnitude-class}
\alias{new_magnitude}
\title{Magnitude class}
\usage{
new_magnitude(dbname, ...)
}
\arguments{
\item{dbname}{the path to the magnitude file.}

\item{...}{other arguments are passed to \code{RSQLite::dbConnect}.}
}
\value{
a Magnitude class object.
}
\description{
Connection object to a magnitude file.
}
\keyword{internal}
