% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude}
\alias{magnitude}
\title{Create a Magnitude connection}
\usage{
magnitude(path, ...)
}
\arguments{
\item{path}{string; path to a magnitude file.}

\item{...}{other arguments are passed to \code{RSQLite::dbConnect}.}
}
\value{
a Magnitude connection object inheriting
SQLiteConnection class from 'RSQLite' package.
}
\description{
Create a Magnitude connection
}
