% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-simil.R
\name{most_similar}
\alias{most_similar}
\title{Order keys by their similarity to key}
\usage{
most_similar(
  conn,
  key,
  q,
  n = 1L,
  normalized = TRUE,
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching", "faith")
)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{key}{string.}

\item{q}{character vector. elements exact same with key will be dropped from result.}

\item{n}{integer.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{method}{string; method to compute similarity.}
}
\value{
a tibble.
}
\description{
Order keys by their similarity to key
}
