% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-simil.R
\name{most_similar_to_given}
\alias{most_similar_to_given}
\title{Order keys by their similarity to key}
\usage{
most_similar_to_given(
  conn,
  key,
  q,
  n = 1L,
  normalized = TRUE,
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching", "faith")
)
}
\arguments{
\item{key}{string.}

\item{q}{character vector.}

\item{n}{integer.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{con}{a Magnitude connection.}

\item{mehtod}{string; method to compute similarity.}
}
\value{
an ordered named numeric vector of which elements represent similarities to `key`.
}
\description{
Order keys by their similarity to key
}
