% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Get vector embeddings of keys}
\usage{
query(
  conn,
  q,
  normalized = TRUE,
  ngram_beg = NULL,
  ngram_end = NULL,
  topn = 3L
)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{q}{a character vector.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{ngram_beg}{integer. If supplied, the function gets out-of-vocabulary vectors
by using character ngrams of which length are `ngram_end - ngram_beg`.}

\item{ngram_end}{integer.}

\item{topn}{integer used for making out-of-vocabulary vectors.}
}
\value{
a tibble.
}
\description{
Get vector embeddings of keys.
If out of vocabulary, their embeddings would be generated at random.
}
