% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_dist}
\alias{query_dist}
\title{Calculates the distance from keys to the keys in q}
\usage{
query_dist(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming"),
  ...
)
}
\arguments{
\item{con}{A Magnitude connection.}

\item{key}{Character vector.}

\item{q}{Character vector.}

\item{normalized}{Logical.}

\item{...}{Other arguments are passed to \code{proxyC::dist}.}

\item{mehtod}{String; method to compute distance.}
}
\value{
A sparse Matrix of 'Matrix' package.
}
\description{
Calculates the distance from keys to the keys in q
}
