% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_simil}
\alias{query_simil}
\title{Calculate the similarity from keys to the keys in q}
\usage{
query_simil(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching", "faith"),
  ...
)
}
\arguments{
\item{con}{A Magnitude connection.}

\item{key}{Character vector.}

\item{q}{Character vector.}

\item{normalized}{Logical.}

\item{...}{Other arguments are passed to \code{proxyC::simil}.}

\item{mehtod}{String; method to compute similarity.}
}
\value{
A sparse Matrix of 'Matrix' package.
}
\description{
Calculate the similarity from keys to the keys in q
}
