% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_frac}
\alias{slice_frac}
\title{Slice samples by fraction from a Magnitude table}
\usage{
slice_frac(conn, frac = 0.001, normalized = TRUE)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{frac}{numeric.}

\item{normalized}{logical;
whether or not vector embeddings should be normalized?}
}
\value{
a tibble.
}
\description{
Slice samples by fraction from a Magnitude table
}
