% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_index}
\alias{slice_index}
\title{Slice samples by index from a Magnitude table}
\usage{
slice_index(conn, index, normalized = TRUE)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{index}{integer vector.}

\item{normalized}{logical;
whether or not vector embeddings should be normalized?}
}
\value{
a tibble.
}
\description{
Slice samples by index from a Magnitude table
}
