% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-wrd.R
\name{wrd}
\alias{wrd}
\title{Calculate Word Rotator's Distance}
\usage{
wrd(x, y, ...)
}
\arguments{
\item{x}{a dense or sparse matrix.}

\item{y}{a dense or sparse matrix.}

\item{...}{other arguments are passed to \code{transport::wasserstein} interenally.}
}
\value{
numeric scalar.
}
\description{
Calculate Word Rotator's Distance between two distributions.
}
\details{
Word Rotator's Distance is a measure of textual similarity
improved of Word Mover's Distance.
}
\seealso{
\url{http://dx.doi.org/10.18653/v1/2020.emnlp-main.236}
}
