% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportita-package.R
\docType{package}
\name{apportita-package}
\alias{apportita}
\alias{apportita-package}
\title{apportita: Utilities for Handling 'magnitude' Word Embeddings}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
