% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distances from keys to keys}
\usage{
calc_dist(
  conn,
  keys,
  q,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming"),
  ...
)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{keys}{character vector.}

\item{q}{character vector.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{method}{string; method to compute distance.}

\item{...}{other arguments are passed to \code{proxyC::dist}.}
}
\value{
a sparse Matrix of 'Matrix' package.
}
\description{
Calculate distances from keys to keys
}
