% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{calc_simil}
\alias{calc_simil}
\title{Calculate similarities from keys to keys in q}
\usage{
calc_simil(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching", "faith"),
  ...
)
}
\arguments{
\item{con}{a Magnitude connection.}

\item{key}{character vector.}

\item{q}{character vector.}

\item{normalized}{logical.}

\item{...}{other arguments are passed to \code{proxyC::simil}.}

\item{mehtod}{string; method to compute similarity.}
}
\value{
a sparse Matrix of 'Matrix' package.
}
\description{
Calculate similarities from keys to keys in q
}
