% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-simil.R
\name{calc_simil}
\alias{calc_simil}
\title{Calculate similarities from keys to keys}
\usage{
calc_simil(
  conn,
  keys,
  q,
  normalized = TRUE,
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching", "faith"),
  ...
)
}
\arguments{
\item{keys}{character vector.}

\item{q}{character vector.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{...}{other arguments are passed to \code{proxyC::simil}.}

\item{con}{a Magnitude connection.}

\item{mehtod}{string; method to compute similarity.}
}
\value{
a sparse Matrix of 'Matrix' package.
}
\description{
Calculate similarities from keys to keys
}
