% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{close}
\alias{close}
\title{Close a Magnitude connection}
\usage{
close(con)
}
\arguments{
\item{con}{A Magnitude connection.}
}
\value{
The value from \code{RSQLite::dbDisconnect} is returned invisibly.
}
\description{
Close a Magnitude connection
}
