% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{doesnt_match}
\alias{doesnt_match}
\title{Find the farest key out of q}
\usage{
doesnt_match(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming"),
  n = 1L
)
}
\arguments{
\item{con}{A Magnitude connection.}

\item{key}{String.}

\item{q}{Character vector.}

\item{normalized}{Logical.}

\item{method}{String; method to compute distance.}

\item{n}{Integer.}
}
\value{
A named numeric vector of which elements represent distances to `key`.
}
\description{
Given a set of keys, figures out which key doesn't match the rest.
}
