% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-dist.R
\name{doesnt_match}
\alias{doesnt_match}
\title{Order keys by their distances to key}
\usage{
doesnt_match(
  conn,
  key,
  q,
  n = 1L,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming")
)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{key}{string.}

\item{q}{character vector.}

\item{n}{integer.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{method}{string; method to compute distance.}
}
\value{
an ordered named numeric vector of which elements represent distances to `key`.
}
\description{
Order keys by their distances to key
}
