% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-dist.R
\name{doesnt_match}
\alias{doesnt_match}
\title{Order keys by their distances to a key}
\usage{
doesnt_match(
  conn,
  key,
  q,
  n = 1L,
  normalized = TRUE,
  method = c("euclidean", "chisquared", "kullback", "manhattan", "maximum", "canberra",
    "minkowski", "hamming")
)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{key}{string.}

\item{q}{character vector. elements exact same with key will be dropped from result.}

\item{n}{integer.}

\item{normalized}{logical; whether or not vector embeddings should be normalized?}

\item{method}{string; method to compute distance.}
}
\value{
a tibble.
}
\description{
Order keys by their distances to a key
}
