% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{most_similar_to_given}
\alias{most_similar_to_given}
\title{Find the most similar key out of q}
\usage{
most_similar_to_given(
  con,
  key,
  q,
  normalized = TRUE,
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching", "faith"),
  n = 1L
)
}
\arguments{
\item{con}{A Magnitude connection.}

\item{key}{String.}

\item{q}{Character vector.}

\item{normalized}{Logical.}

\item{n}{Integer.}

\item{mehtod}{String; method to compute similarity.}
}
\value{
A named numeric vector of which elements represent similarities to `key`.
}
\description{
Query for the most similar key out of a list of keys
to a given key like so.
}
