% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(con, q, normalized = TRUE)
}
\arguments{
\item{con}{A Magnitude connection.}

\item{q}{Character vector;}

\item{normalized}{Logical;}
}
\value{
A tibble.
}
\description{
Handles a query of keys which could be a list (a character vector)
of keys.
}
