% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_frac}
\alias{slice_frac}
\title{Slice samples by frac from a Magnitude table}
\usage{
slice_frac(con, frac = 0.001, normalized = TRUE)
}
\arguments{
\item{con}{a Magnitude connection.}

\item{frac}{numeric.}

\item{normalized}{logical;}
}
\value{
a tibble.
}
\description{
Slice samples by frac from a Magnitude table
}
