% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_n}
\alias{slice_n}
\title{Slice}
\usage{
slice_n(con, n, offset = 0, normalized = FALSE)
}
\arguments{
\item{con}{A Magnitude connection.}

\item{n}{Integer.}

\item{offset}{Integer.}

\item{normalized}{Logical;}
}
\value{
A tibble.
}
\description{
Slice
}
