% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_n}
\alias{slice_n}
\title{Slice samples from a Magnitude table}
\usage{
slice_n(conn, n, offset = 0, normalized = TRUE)
}
\arguments{
\item{conn}{a Magnitude connection.}

\item{n}{integer.}

\item{offset}{integer.}

\item{normalized}{logical;
whether or not vector embeddings should be normalized?}
}
\value{
a tibble.
}
\description{
Slice samples from a Magnitude table
}
